fix(build): parse markdown with mdast-util-from-markdown, not unified (#577 build break)
The merged #577 broke the browser builds (vscode webview + viewer app):
`npm run build` failed with "path/process/url has been externalized" and
"fileURLToPath is not exported". Root cause: `getMarkdownImageReferences`
parsed via `unified().use(remarkParse)`, which drags in `vfile` — whose
`minurl` module imports Node built-ins. That function is re-exported from the
`stagebook` barrel, which the browser-facing `stagebook` / `stagebook/viewer`
entries transitively import, and tsup bundled `vfile` into a shared chunk the
viewer pulls (alongside `fillTemplates`). CI on the feature branch caught it;
#577's own CI didn't, because a PR based on a feature branch gets no CI.

Fix: parse with `mdast-util-from-markdown` — the low-level CommonMark parser
that `remark-parse` wraps and that `react-markdown` itself uses browser-side.
It yields the IDENTICAL mdast (same image/imageReference nodes, urls, alt,
positions — verified) with NO `vfile` / Node built-ins, so it bundles cleanly
for every consumer. `unified` + `remark-parse` are dropped as direct deps;
`mdast-util-from-markdown` is added and `noExternal`-bundled for the CJS build.

Also splits the parser into its own module (`markdownImageReferences.ts` +
test) so bundles that never enumerate markdown can tree-shake it, and exports
the shared `isCollectableLocalPath` for it.

Full root build, both browser bundles, `require()`/`import()`, and all tests
(stagebook 2127 / viewer 79 / vscode 157) pass.
