+++ packages/stagebook/package.json
@@ -98,11 +98,10 @@
   "dependencies": {
     "@hello-pangea/dnd": "^18.0.1",
     "js-yaml": "^4.1.0",
+    "mdast-util-from-markdown": "^2.0.0",
     "react-markdown": "^10.1.0",
     "remark-gfm": "^4.0.1",
-    "remark-parse": "^11.0.0",
     "tinyglobby": "^0.2.16",
-    "unified": "^11.0.0",
     "yaml": "^2.8.3"
   },
   "peerDependencies": {
